# lib/openapi/spec.json
{
  "openapi": "3.0.0",
  "info": {
    "title": "Timesheet API",
    "version": "1.0.0",
    "description": "Complete REST API for the Timesheet workforce management system. Handles employee time tracking, scheduling, roster management, payroll and analytics.\n\n## Authentication\nUses cookie-based JWT authentication. Three separate auth layers:\n- Admin/user: login via /api/auth/login\n- Employee: login via /api/employee/login\n- Device: register via /api/device/register\n\n## Roles\n- super_admin: full system access\n- admin: manage employees and schedules\n- user: limited view-only dashboard access\n- employee: clock in/out and own timesheets\n\n## Base URL\nAll endpoints are prefixed with /api\n\n## Error Responses\nAll errors follow the format: { error: string, message?: string }\n\n## Cookie Auth\nAll protected endpoints require valid JWT cookie sent automatically via credentials: include on all requests."
  },
  "servers": [
    {
      "url": "/api",
      "description": "API Server"
    }
  ],
  "components": {
    "securitySchemes": {
      "cookieAuth": {
        "type": "apiKey",
        "in": "cookie",
        "name": "session"
      }
    },
    "schemas": {
      "Error": {
        "type": "object",
        "properties": {
          "error": {
            "type": "string",
            "example": "Validation failed"
          },
          "message": {
            "type": "string",
            "example": "Invalid input data"
          },
          "details": {
            "type": "object"
          }
        },
        "required": ["error"]
      },
...
      "ObjectId": {
        "type": "string",
        "pattern": "^[0-9a-fA-F]{24}$",
        "example": "507f1f77bcf86cd799439011"
      },
      "DateTime": {
        "type": "string",
        "format": "date-time",
        "example": "2024-03-07T10:30:00.000Z"
      },
      "Date": {
        "type": "string",
        "format": "date",
        "example": "2024-03-07"
      },
...
      "Timesheet": {
        "type": "object",
        "properties": {
          "id": {
            "$ref": "#/components/schemas/ObjectId"
          },
          "employeeId": {
            "$ref": "#/components/schemas/ObjectId"
          },
          "date": {
            "$ref": "#/components/schemas/Date"
          },
          "clockIn": {
            "$ref": "#/components/schemas/DateTime"
          },
          "clockOut": {
            "$ref": "#/components/schemas/DateTime"
          },
          "breakDuration": {
            "type": "number",
            "minimum": 0,
            "example": 30
          },
          "totalHours": {
            "type": "number",
            "example": 8.5
          },
          "notes": {
            "type": "string",
            "example": "Worked overtime on project deadline"
          },
          "locationId": {
            "$ref": "#/components/schemas/ObjectId"
          },
          "shiftId": {
            "$ref": "#/components/schemas/ObjectId"
          },
          "createdAt": {
            "$ref": "#/components/schemas/DateTime"
          },
          "updatedAt": {
            "$ref": "#/components/schemas/DateTime"
          }
        }
      },
...
    }
  },
  "security": [
    {
      "cookieAuth": []
    }
  ],
  "paths": {
...
    "/employees/{id}/timesheet": {
      "get": {
        "tags": ["Employee Timesheets"],
        "summary": "Get employee timesheet entries with date filtering",
        "description": "Retrieves timesheet entries for a specific employee within the specified date range. Returns clock-in/out times, break durations, total hours worked, and any associated notes or flags. Supports pagination and date filtering. Accessible by admin users and the employee themselves. Used for payroll processing and attendance tracking.",
        "parameters": [
          {
            "name": "id",
            "in": "path",
            "required": true,
            "schema": {
              "$ref": "#/components/schemas/ObjectId"
            }
          },
          {
            "name": "startDate",
            "in": "query",
            "schema": {
              "$ref": "#/components/schemas/Date"
            }
          },
          {
            "name": "endDate",
            "in": "query",
            "schema": {
              "$ref": "#/components/schemas/Date"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Employee timesheet data",
            "content": {
              "application/json": {
                "schema": {
                  "type": "array",
                  "items": {
                    "$ref": "#/components/schemas/Timesheet"
                  }
                }
              }
            }
          },
          "401": {
            "description": "Unauthorized",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/Error"
                }
              }
            }
          },
          "404": {
            "description": "Employee not found",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/Error"
                }
              }
            }
          },
          "500": {
            "description": "Internal server error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/Error"
                }
              }
            }
          }
        }
      },
      "patch": {
        "tags": ["Employee Timesheets"],
        "summary": "Update employee timesheet entry",
        "description": "Allows admin users to modify timesheet entries for corrections or adjustments. Can update clock-out times, break durations, and add administrative notes. Automatically recalculates total hours worked. Logs all changes for audit purposes. Requires admin or super_admin role. Used for correcting time tracking errors or manual adjustments.",
        "parameters": [
          {
            "name": "id",
            "in": "path",
            "required": true,
            "schema": {
              "$ref": "#/components/schemas/ObjectId"
            }
          }
        ],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "type": "object",
                "properties": {
                  "clockOut": {
                    "$ref": "#/components/schemas/DateTime"
                  },
                  "breakDuration": {
                    "type": "number",
                    "minimum": 0
                  },
                  "notes": {
                    "type": "string",
                    "maxLength": 500
                  }
                }
              }
            }
          }
        },
        "responses": {
          "200": {
            "description": "Timesheet updated successfully",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/Timesheet"
                }
              }
            }
          },
          "400": {
            "description": "Invalid request data",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/Error"
                }
              }
            }
          },
          "401": {
            "description": "Unauthorized",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/Error"
                }
              }
            }
          },
          "404": {
            "description": "Employee not found",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/Error"
                }
              }
            }
          },
          "500": {
            "description": "Internal server error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/Error"
                }
              }
            }
          }
        }
      }
    },
...
    "/employees/check-pin": {
      "get": {
        "tags": ["Employees"],
        "summary": "Check if employee PIN already exists in system",
        "description": "Validates whether a specific PIN is already assigned to an employee to prevent duplicates. Returns existence status and associated employee ID if found. Used during employee creation and PIN updates to ensure uniqueness. Essential for maintaining security and preventing PIN conflicts in the time clock system. Accessible by admin users during employee management.",
        "parameters": [
          {
            "name": "pin",
            "in": "query",
            "required": true,
            "schema": {
              "type": "string",
              "example": "1234"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "PIN check result",
            "content": {
              "application/json": {
                "schema": {
                  "type": "object",
                  "properties": {
                    "exists": {
                      "type": "boolean",
                      "example": true
                    },
                    "employeeId": {
                      "$ref": "#/components/schemas/ObjectId"
                    }
                  }
                }
              }
            }
          },
          "400": {
            "description": "Invalid PIN format",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/Error"
                }
              }
            }
          },
          "401": {
            "description": "Unauthorized",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/Error"
                }
              }
            }
          },
          "500": {
            "description": "Internal server error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/Error"
                }
              }
            }
          }
        }
      }
    },
...
  }
}